Update instance.pp
composer-example.json only exists for MW < 1.25

+++ manifests/instance.pp
@@ -110,9 +110,11 @@ define mediawiki::instance (
         group   => $apache_group,
       }
       
+      # Only for MW < 1.25
       exec { "${name}-composer.json":
         command => "cp ${doc_root}/${name}/composer-example.json ${doc_root}/${name}/composer.json",
         path    => "/usr/local/bin/:/bin/",
+        onlyif  => "test -f ${doc_root}/${name}/composer-example.json",
       }
       File["${doc_root}/${name}"] -> Exec["${name}-install_script"]
       Exec["${name}-install_script"] -> Exec["${name}-composer.json"]
